Refresh Java CompletableFuture shadow evidence
