+++ bench/recall_loss/scheduling-lifecycle-boundary-audit-java-completablefuture-constructor-reporting-2026-07-02.v1.json
@@ -82,9 +82,9 @@
       "admission_rejections": 812,
       "canon_checked": 99,
       "canon_preservation_violations": 0,
-      "excluded_units": 6003,
+      "excluded_units": 6011,
       "interpretable_units": 970,
-      "total_units": 6973
+      "total_units": 6981
     }
   },
   "generated_on": "2026-07-02",
