+++ docs/recall-loss-diagnostics.md
@@ -202,7 +202,7 @@ reporting-supported closed-boundary status and leaves `230` broad
 `CompletableFuture` mentions closed for later splitting. The Java-heavy query
 regression kept product output hashes identical across `netty`, `rxjava`,
 `retrofit`, `junit5`, `jedis`, and `h2database`, with aggregate median runtime
-`7387.63ms -> 7366.03ms` (`-0.29%`). The
+`7108.09ms -> 7188.73ms` (`+1.13%`). The
 follow-up [Rust block_on future-drive artifact](../bench/recall_loss/rust-block-on-future-drive-obligation-reporting-2026-07-01.v1.json)
 keeps exact admission closed and maps qualified/import-backed Rust
 `tokio_test::block_on` calls plus proof-backed tokio runtime receiver chains to
